Fix loop index type casting in SnakeGapSbmProcess for consistency

+++ applications/IgaApplication/custom_processes/snake_gap_sbm_process.h
@@ -116,7 +116,7 @@ public:
         const auto first = ci.begin() + b;
         const auto last  = ci.begin() + e;
         const auto it = std::lower_bound(first, last, j);
-        if (it != last && *it == j) return static_cast<std::size_t>(it - ci.begin());
+        if (it != last && static_cast<std::size_t>(*it) == j) return static_cast<std::size_t>(it - ci.begin());
         return static_cast<std::size_t>(-1);
     }
